pyproject.toml
[tool.poetry]
name = "pymine"
version = "0.1.0"
description = "Python bindings for Minecraft Education Edition"
authors = ["Ollie Cheng <denosawr@vivaldi.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.1"
websockets = "^8.1"
pycryptodome = "^3.9.8"
colorlog = "^4.2.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
server = "pymine.server:start_server"
client = "pymine.client:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[[tool.poetry.source]]
name = "pypi_with_cert"
url = "https://pypi.org/project/"
default = true
